refactor: create separate folder for external/shared traits
