+++ azns_registry/Cargo.toml
@@ -8,6 +8,7 @@ publish = false
 [dependencies]
 ink = { version = "4.0.0", default-features = false }
 
+interfaces = { path = "../interfaces", default-features = false }
 azns_name_checker = { path = "../azns_name_checker", default-features = false, features = ["ink-as-dependency"] }
 azns_fee_calculator = { path = "../azns_fee_calculator", default-features = false, features = ["ink-as-dependency"] }
 azns_merkle_verifier = { path = "../azns_merkle_verifier", default-features = false, features = ["ink-as-dependency"] }
@@ -30,6 +31,7 @@ std = [
     "ink/std",
     "scale/std",
     "scale-info/std",
+    "interfaces/std",
     "azns_name_checker/std",
     "azns_fee_calculator/std",
     "azns_merkle_verifier/std",
