+++ azns_registry/lib.rs
@@ -1,20 +1,18 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 mod address_dict;
-mod art_zero_traits;
-mod psp34_standard;
 
 #[ink::contract]
 mod azns_registry {
     use crate::address_dict::AddressDict;
-    use crate::art_zero_traits::*;
-    use crate::psp34_standard::*;
     use ink::env::call::FromAccountId;
     use ink::env::hash::CryptoHash;
     use ink::prelude::string::{String, ToString};
     use ink::prelude::vec::Vec;
     use ink::storage::traits::ManualKey;
     use ink::storage::{Lazy, Mapping};
+    use interfaces::art_zero_traits::*;
+    use interfaces::psp34_standard::*;
 
     use azns_fee_calculator::FeeCalculatorRef;
     use azns_merkle_verifier::MerkleVerifierRef;
