+++ interfaces/lib.rs
@@ -1,0 +1,4 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+
+pub mod art_zero_traits;
+pub mod psp34_standard;
